dev: Move cell layout emitter into a submodule
